+++ uniservice_lib/__init__.py
@@ -21,4 +21,4 @@ from __future__ import annotations
 
 __all__ = ["__version__"]
 
-__version__ = "1.2.0"
+__version__ = "1.2.1"